[#53] refactor: CategoryMapService 모듈 이동 (client -> common)
