+++ kyumall-client/src/main/java/com/kyumall/kyumallclient/product/ProductService.java
@@ -11,6 +11,7 @@ import com.kyumall.kyumallcommon.product.entity.Category;
 import com.kyumall.kyumallcommon.product.entity.Product;
 import com.kyumall.kyumallcommon.product.repository.CategoryRepository;
 import com.kyumall.kyumallcommon.product.repository.ProductRepository;
+import com.kyumall.kyumallcommon.product.service.CategoryMapService;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
